requirements.txt
astroid==2.1.0
colorama==0.4.1
Django==2.1.4
django-crispy-forms==1.7.2
django-filter==2.0.0
django-pagedown==1.0.6
djangorestframework==3.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
psycopg2==2.7.6.1
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.7
six==1.12.0
typed-ast==1.1.1
wrapt==1.10.11
